+++ BACtrack/app/src/main/AndroidManifest.xml
@@ -13,6 +13,11 @@
         android:theme="@style/Theme.BACtrack"
         tools:targetApi="31">
         <activity
+            android:name=".Landing"
+            android:exported="false"
+            android:label="@string/title_activity_landing"
+            android:theme="@style/Theme.BACtrack" />
+        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
