+++ BACtrack/app/src/main/java/com/example/bactrack/Landing.kt
@@ -1,0 +1,548 @@
+package com.example.bactrack
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+
+import androidx.compose.foundation.lazy.LazyColumn
+
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.text.KeyboardOptions
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Email
+import androidx.compose.material.icons.filled.Face
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.icons.filled.SportsBar
+import androidx.compose.material.icons.outlined.Email
+import androidx.compose.material.icons.outlined.Face
+import androidx.compose.material.icons.outlined.Home
+import androidx.compose.material.icons.outlined.Settings
+import androidx.compose.material.icons.outlined.SportsBar
+import androidx.compose.material3.Badge
+import androidx.compose.material3.BadgedBox
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+
+import androidx.compose.ui.res.painterResource
+
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.input.KeyboardType
+
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.navigation.compose.rememberNavController
+import com.example.bactrack.ui.theme.BACtrackTheme
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.googlefonts.Font
+import androidx.compose.ui.text.googlefonts.GoogleFont
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.bactrack.uiScreens.DisplayOne
+
+
+class Landing : ComponentActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            BACtrackTheme {
+                val items = listOf(
+                    BottomNavigationItem(
+                        title = "Home",
+                        selectedIcon = Icons.Filled.Home,
+                        unselectedIcon = Icons.Outlined.Home,
+                        hasNews = false
+                    ),
+                    BottomNavigationItem(
+                        title = "Your History",
+                        selectedIcon = Icons.Filled.SportsBar,
+                        unselectedIcon = Icons.Outlined.SportsBar,
+                        hasNews = false
+                    ),
+                    BottomNavigationItem(
+                        title = "Settings",
+                        selectedIcon = Icons.Filled.Settings,
+                        unselectedIcon = Icons.Outlined.Settings,
+                        hasNews = false
+                    ),
+                    BottomNavigationItem(
+                        title = "Profile",
+                        selectedIcon = Icons.Filled.Face,  // Choose an appropriate icon for profile
+                        unselectedIcon = Icons.Outlined.Face,
+                        hasNews = false
+                    )
+                )
+                var selectedItemIndex by rememberSaveable { mutableStateOf(0) }
+
+                var navController = rememberNavController()
+
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = Color(0xFFADD8E6) //color can be subject to change
+                ) {
+                    Scaffold(
+                        bottomBar = {
+                            BottomNavigationBar(
+                                items = items,
+                                selectedItemIndex = selectedItemIndex,
+                                onItemSelected = { index ->
+                                    selectedItemIndex = index
+                                    when(index) {
+                                        0 -> navController.navigate("home")
+                                        1 -> navController.navigate("history")
+                                        2 -> navController.navigate("settings")
+                                        3 -> navController.navigate("profile")
+                                    }
+                                }
+                            )
+                        }
+                    ) { paddingValues -> // This will ensure our main content is not covered by the bottom bar
+                        NavHost(
+                            navController = navController,
+                            startDestination = "home",
+                            modifier = Modifier.padding(paddingValues)
+                        ) {
+                            composable("home") { HomeScreen() }
+                            composable("history") { HistoryScreen() }
+                            composable("settings") { SettingsScreen() }
+                            composable("profile") { ProfileMenu() }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+data class BottomNavigationItem (
+    val title: String,
+    val selectedIcon: ImageVector,
+    val unselectedIcon: ImageVector,
+    //this will represent a number of notifications the icon will contain (optional)
+    val hasNews: Boolean,
+    val badgeCount: Int? = null
+)
+
+@Composable
+fun Greeting(name: String, modifier: Modifier = Modifier) {
+    Text(
+        text = "Hello $name!",
+        modifier = modifier
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun GreetingPreview() {
+    BACtrackTheme {
+        Greeting("Android")
+    }
+}
+
+@Composable
+fun BottomNavigationBar(
+    items: List<BottomNavigationItem>,
+    selectedItemIndex: Int,
+    onItemSelected: (Int) -> Unit
+) {
+    NavigationBar {
+        items.forEachIndexed { index, item ->
+            NavigationBarItem(
+                selected = selectedItemIndex == index,
+                onClick = { onItemSelected(index)},
+                label = { Text(text = item.title) },
+                icon = {
+                    BadgedBox( ////////////////this is boiler plate code for the event we want to add notifications to the icons
+                        badge = {
+                            if(item.badgeCount != null) {
+                                Badge{ Text(text = item.badgeCount.toString()) }
+                            } else if(item.hasNews) { Badge()}
+                        }
+                    ) { ///////////////////////
+                        Icon(
+                            imageVector = if(index == selectedItemIndex) {
+                                item.selectedIcon
+                            } else {
+                                item.unselectedIcon
+                            },
+                            contentDescription = item.title
+                        )
+                    }
+                }
+            )
+
+        }
+    }
+}
+
+@Composable
+fun HomeScreen() {
+    var counter by remember { mutableStateOf(0) }
+    val maxCounter = 10 // Set a max level for the mug
+    val fillLevel by animateFloatAsState(targetValue = (counter / maxCounter.toFloat()).coerceIn(0f, 1f))
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = Color(0xFFADD8E6)
+
+    ) {
+        Box(
+            contentAlignment = Alignment.Center,
+            modifier = Modifier.fillMaxSize()
+        ) {
+            LazyColumn(
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                item {
+                    Image(
+                        painter = painterResource(id = R.drawable.bactrack_logo_nobackround),
+                        contentDescription = "BACtrack Logo",
+                        modifier = Modifier
+                            .size(200.dp)
+                            .padding(top = 75.dp)
+
+                    )
+                }
+                item {
+
+                    Text(
+                        text = "Welcome to \n \n BACtrack",
+
+                        color = Color.Black,
+                        fontWeight = FontWeight.Bold,
+                        style = MaterialTheme.typography.bodyLarge,
+                        fontSize = 36.sp
+                    )
+                }
+                item {
+                    DisplayOne() //This shows the Add a drink button
+                }
+
+                item {
+                    Mug(fillLevel = fillLevel)
+
+                    Spacer(modifier = Modifier.height(16.dp))
+
+                    Button(onClick = { if (counter < maxCounter) counter++ }) {
+                        Text(text = "Increase Count")
+                    }
+                    Button(onClick = { if (counter < maxCounter) counter-- }) {
+                        Text(text = "Decrease Count")
+                    }
+                }
+
+                item {
+                    Text(
+                        modifier = Modifier.padding(top = 1.dp),
+                        text = "Your BAC is : " + SessionManager.totalDrinks.toString(),
+                        color = Color.Red,
+                        fontWeight = FontWeight.Bold,
+                        style = MaterialTheme.typography.bodyLarge,
+                        fontSize = 48.sp
+                    )
+                }
+
+            }
+
+        }
+    }
+}
+
+@Composable
+fun Mug(fillLevel: Float) {
+    val mugWidth = 100.dp
+    val mugHeight = 150.dp
+
+    Canvas(modifier = Modifier.size(mugWidth, mugHeight)) {
+        // Draw the mug outline
+        drawRoundRect(
+            color = Color.Gray,
+            size = size,
+            style = Stroke(width = 8f)
+        )
+
+        // Draw the filling (based on fill level, starting from the bottom)
+        drawRect(
+            color = Color.Blue,
+            size = size.copy(height = size.height * fillLevel),
+            topLeft = androidx.compose.ui.geometry.Offset(
+                x = 0f,
+                y = size.height * (1 - fillLevel)  // Starts from the bottom and goes upwards
+            )
+        )
+    }
+}
+
+@Composable
+fun HistoryScreen() {
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = Color(0xFFADD8E6)
+    ) {
+        Text(
+            text = "History Screen",
+            color = Color.Black
+        )
+    }
+}
+
+@Composable
+fun SettingsScreen() {
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = Color(0xFFADD8E6)
+    ) {
+        Text(
+            text = "Settings Screen",
+            color = Color.Black
+        )
+    }
+}
+
+
+data class CurrentSession(
+    val numBeers: Int = 0,
+    val numWine: Int = 0,
+    val numShots: Int = 0,
+    val numCocktails: Int = 0
+) {
+
+}
+
+
+@Composable
+fun ProfileMenu() {
+    val menuItems = listOf("Personal Info", "Body Measurements", "Preferences")
+    var selectedMenuItem by remember { mutableStateOf("Personal Info") }
+
+    Row(modifier = Modifier.fillMaxSize()) {
+        // Menu Column on the left
+        Column(
+            modifier = Modifier
+                .fillMaxHeight()
+                .padding(16.dp)
+                .width(150.dp)
+        ) {
+            menuItems.forEach { item ->
+                Button(
+                    onClick = { selectedMenuItem = item },
+                    modifier = Modifier.fillMaxWidth(),
+                    colors = ButtonDefaults.buttonColors(
+                        containerColor = if (selectedMenuItem == item) Color.Gray else Color.LightGray
+                    )
+                ) {
+                    Text(text = item)
+                }
+                Spacer(modifier = Modifier.height(8.dp))
+            }
+        }
+
+        // Dynamic Content on the right
+        Column(
+            modifier = Modifier
+                .fillMaxHeight()
+                .padding(16.dp)
+                .weight(1f)
+        ) {
+            when (selectedMenuItem) {
+                "Personal Info" -> PersonalInfoTab()
+//                "Profile Picture" -> ProfilePictureTab()
+                "Body Measurements" -> BodyMeasurementsTab()
+                "Preferences" -> PreferencesTab()
+            }
+        }
+    }
+}
+
+@SuppressLint("RememberReturnType")
+@Composable
+fun PersonalInfoTab() {
+    val context = LocalContext.current
+    val sharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
+    var name by remember { mutableStateOf(sharedPreferences.getString("name", "") ?: "") }
+    var gender by remember { mutableStateOf(sharedPreferences.getString("gender", "") ?: "") }
+
+
+
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = Color(0xFFADD8E6)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            // Display profile picture (placeholder image for now)
+            Image(
+                painter = painterResource(id = R.drawable.profile_user), // Use a drawable resource
+                contentDescription = "Profile Picture",
+                modifier = Modifier
+                    .height(100.dp)
+                    .width(100.dp)
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            // Display name information
+            Text(
+                text = "Name: $name",
+                style = MaterialTheme.typography.titleLarge,  // Updated text style ,
+                modifier = Modifier.fillMaxWidth(),
+                color = Color.Black
+            )
+
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+//
+            val gender = "Male"
+            Text(
+                text = "Gender: $gender",
+                style = MaterialTheme.typography.titleLarge,  // Use titleLarge instead of h6
+                modifier = Modifier.fillMaxWidth(),
+                color = Color.Black
+            )
+
+            Button(onClick = {
+                // Perform actions to navigate or edit profile info
+            }) {
+                Text(text = "Edit Profile Info")
+            }
+        }
+    }
+}
+
+@Composable
+fun BodyMeasurementsTab() {
+    val context = LocalContext.current
+    val sharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
+    var weight by remember { mutableStateOf(sharedPreferences.getString("weight", "") ?: "") }
+    var height by remember { mutableStateOf(sharedPreferences.getString("height", "") ?: "") }
+    var showMessage by remember { mutableStateOf(false) }
+
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = Color(0xFFADD8E6)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            OutlinedTextField(
+                value = weight,
+                onValueChange = { weight = it },
+                label = { Text("Enter your weight (kg)") },
+                modifier = Modifier.fillMaxWidth(),
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            OutlinedTextField(
+                value = height,
+                onValueChange = { height = it },
+                label = { Text("Enter your height (cm)") },
+                modifier = Modifier.fillMaxWidth(),
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Button(onClick = {
+                if (weight.isNotBlank() && height.isNotBlank()) {
+                    val editor = sharedPreferences.edit()
+                    editor.putString("weight", weight)
+                    editor.putString("height", height)
+                    editor.apply()
+                    showMessage = true
+                }
+            }) {
+                Text(text = "Save Measurements")
+            }
+
+            if (showMessage) {
+                Text(
+                    text = "Measurements saved successfully!",
+                    color = Color.Green,
+                    modifier = Modifier.padding(top = 16.dp)
+                )
+            }
+        }
+    }
+}
+
+@Composable
+fun PreferencesTab() {
+    // Placeholder for Preferences functionality
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = Color(0xFFADD8E6)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text("Preferences Section")
+
+            // This is where you'd add preference settings like notifications, themes, etc.
+            Button(onClick = { /* Implement preferences logic */ }) {
+                Text(text = "Edit Preferences")
+            }
+        }
+    }
+}
