+++ BACtrack/app/src/main/java/com/example/bactrack/MainActivity.kt
@@ -1,518 +1,190 @@
 package com.example.bactrack
 
-import android.annotation.SuppressLint
-import android.content.Context
+
+import android.content.Intent
 import android.os.Bundle
+
+import android.util.Log
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-
-import androidx.compose.foundation.lazy.LazyColumn
-
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.text.KeyboardOptions
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.Face
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.filled.SportsBar
-import androidx.compose.material.icons.outlined.Email
-import androidx.compose.material.icons.outlined.Face
-import androidx.compose.material.icons.outlined.Home
-import androidx.compose.material.icons.outlined.Settings
-import androidx.compose.material.icons.outlined.SportsBar
-import androidx.compose.material3.Badge
-import androidx.compose.material3.BadgedBox
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-
-import androidx.compose.ui.res.painterResource
-
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.input.KeyboardType
-
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.navigation.compose.rememberNavController
 import com.example.bactrack.ui.theme.BACtrackTheme
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.compose.foundation.Image
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.FirebaseApp
+import android.widget.Toast
+import com.google.firebase.auth.OAuthProvider
+import androidx.activity.compose.setContent
+
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.googlefonts.Font
-import androidx.compose.ui.text.googlefonts.GoogleFont
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.example.bactrack.SessionManager.totalAlcMass
-import com.example.bactrack.uiScreens.DisplayOne
-
+import androidx.credentials.CredentialManager
+import androidx.credentials.GetCredentialRequest
+import androidx.credentials.exceptions.GetCredentialException
+import com.google.android.libraries.identity.googleid.GetGoogleIdOption
+import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
+import com.google.android.libraries.identity.googleid.GoogleIdTokenParsingException
+import kotlinx.coroutines.launch
+import java.security.MessageDigest
+import java.util.UUID
 
 class MainActivity : ComponentActivity() {
-
+    private lateinit var auth: FirebaseAuth
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
+        FirebaseApp.initializeApp(this)
+        // Get the FirebaseAuth instance
+        auth = FirebaseAuth.getInstance()
+
         setContent {
             BACtrackTheme {
-                val items = listOf(
-                    BottomNavigationItem(
-                        title = "Home",
-                        selectedIcon = Icons.Filled.Home,
-                        unselectedIcon = Icons.Outlined.Home,
-                        hasNews = false
-                    ),
-                    BottomNavigationItem(
-                        title = "Your History",
-                        selectedIcon = Icons.Filled.SportsBar,
-                        unselectedIcon = Icons.Outlined.SportsBar,
-                        hasNews = false
-                    ),
-                    BottomNavigationItem(
-                        title = "Settings",
-                        selectedIcon = Icons.Filled.Settings,
-                        unselectedIcon = Icons.Outlined.Settings,
-                        hasNews = false
-                    ),
-                    BottomNavigationItem(
-                        title = "Profile",
-                        selectedIcon = Icons.Filled.Face,  // Choose an appropriate icon for profile
-                        unselectedIcon = Icons.Outlined.Face,
-                        hasNews = false
-                    )
-                )
-                var selectedItemIndex by rememberSaveable { mutableStateOf(0) }
-
-                var navController = rememberNavController()
-
                 Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = Color(0xFFADD8E6) //color can be subject to change
+                    modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background,
                 ) {
-                    Scaffold(
-                        bottomBar = {
-                            BottomNavigationBar(
-                                items = items,
-                                selectedItemIndex = selectedItemIndex,
-                                onItemSelected = { index ->
-                                    selectedItemIndex = index
-                                    when(index) {
-                                        0 -> navController.navigate("home")
-                                        1 -> navController.navigate("history")
-                                        2 -> navController.navigate("settings")
-                                        3 -> navController.navigate("profile")
-                                    }
-                                }
-                            )
-                        }
-                    ) { paddingValues -> // This will ensure our main content is not covered by the bottom bar
-                        NavHost(
-                            navController = navController,
-                            startDestination = "home",
-                            modifier = Modifier.padding(paddingValues)
-                        ) {
-                            composable("home") { HomeScreen() }
-                            composable("history") { HistoryScreen() }
-                            composable("settings") { SettingsScreen() }
-                            composable("profile") { ProfileMenu() }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-data class BottomNavigationItem (
-    val title: String,
-    val selectedIcon: ImageVector,
-    val unselectedIcon: ImageVector,
-    //this will represent a number of notifications the icon will contain (optional)
-    val hasNews: Boolean,
-    val badgeCount: Int? = null
-)
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    BACtrackTheme {
-        Greeting("Android")
-    }
-}
-
-@Composable
-fun BottomNavigationBar(
-    items: List<BottomNavigationItem>,
-    selectedItemIndex: Int,
-    onItemSelected: (Int) -> Unit
-) {
-    NavigationBar {
-        items.forEachIndexed { index, item ->
-            NavigationBarItem(
-                selected = selectedItemIndex == index,
-                onClick = { onItemSelected(index)},
-                label = { Text(text = item.title) },
-                icon = {
-                    BadgedBox( ////////////////this is boiler plate code for the event we want to add notifications to the icons
-                        badge = {
-                            if(item.badgeCount != null) {
-                                Badge{ Text(text = item.badgeCount.toString()) }
-                            } else if(item.hasNews) { Badge()}
-                        }
-                    ) { ///////////////////////
-                        Icon(
-                            imageVector = if(index == selectedItemIndex) {
-                                item.selectedIcon
-                            } else {
-                                item.unselectedIcon
-                            },
-                            contentDescription = item.title
-                        )
-                    }
-                }
-            )
-
-        }
-    }
-}
-
-@Composable
-fun HomeScreen() {
-
-    // Values for testing.. this will be fixed after MVP!
-    val userWeight = 70.0 // Example weight in kg
-    val userSex = "male" // Example sex
-    val totalAlcoholConsumed = totalAlcMass
-    val timeSinceDrinking = 1.0 // Example: 2 hours since drinking started
-
-
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Color(0xFFADD8E6)
-
-    ) {
-        LazyColumn(
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            item {
-                Image(
-                    painter = painterResource(id = R.drawable.bactrack_logo_nobackround),
-                    contentDescription = "BACtrack Logo",
-                    modifier = Modifier
-                        .size(200.dp)
-                        .padding(top = 75.dp)
-
-                )
-            }
-            item {
-                Text(
-                    text = "Welcome to BACtrack",
-                    color = Color.Black,
-                    fontWeight = FontWeight.Bold,
-                    style = MaterialTheme.typography.bodyLarge,
-                    fontSize = 36.sp
-                )
-            }
-            item {
-                DisplayOne() //This shows the Add a drink button
-            }
-            item {
-                Text(
-                    modifier = Modifier.padding(top = 220.dp),
-                    text = "Your drink count is : " + SessionManager.totalDrinks.toString(),
-                    color = Color.Red,
-                    fontWeight = FontWeight.Bold,
-                    style = MaterialTheme.typography.bodyLarge,
-                    fontSize = 35.sp
-                )
-            }
-
-            item {
-                Text(
-                    modifier = Modifier.padding(top = 100.dp),
-                    text = "Your BAC is : ${calculateBAC(totalAlcoholConsumed, userWeight, userSex, timeSinceDrinking).format(3)}",
-                    color = Color.Red,
-                    fontWeight = FontWeight.Bold,
-                    style = MaterialTheme.typography.bodyLarge,
-                    fontSize = 30.sp
-                )
-            }
-
-        }
-
-    }
-}
-
-fun Double.format(digits: Int) = "%.${digits}f".format(this)
-
-@Composable
-fun HistoryScreen() {
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Color(0xFFADD8E6)
-    ) {
-        Text(
-            text = "History Screen",
-            color = Color.Black
-        )
-    }
-}
-
-@Composable
-fun SettingsScreen() {
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Color(0xFFADD8E6)
-    ) {
-        Text(
-            text = "Settings Screen",
-            color = Color.Black
-        )
-    }
-}
-
-
-data class CurrentSession(
-    val numBeers: Int = 0,
-    val numWine: Int = 0,
-    val numShots: Int = 0,
-    val numCocktails: Int = 0
-) {
-
-}
-
-
-@Composable
-fun ProfileMenu() {
-    val menuItems = listOf("Personal Info", "Body Measurements", "Preferences")
-    var selectedMenuItem by remember { mutableStateOf("Personal Info") }
-
-    Row(modifier = Modifier.fillMaxSize()) {
-        // Menu Column on the left
-        Column(
-            modifier = Modifier
-                .fillMaxHeight()
-                .padding(16.dp)
-                .width(150.dp)
-        ) {
-            menuItems.forEach { item ->
-                Button(
-                    onClick = { selectedMenuItem = item },
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = ButtonDefaults.buttonColors(
-                        containerColor = if (selectedMenuItem == item) Color.Gray else Color.LightGray
+                    Column(
+                        verticalArrangement = Arrangement.Center,
+                        horizontalAlignment = Alignment.CenterHorizontally
                     )
-                ) {
-                    Text(text = item)
+                    {
+
+                        GoogleSignInButton()
+                        GithubButton()
+                    }
                 }
-                Spacer(modifier = Modifier.height(8.dp))
-            }
-        }
-
-        // Dynamic Content on the right
-        Column(
-            modifier = Modifier
-                .fillMaxHeight()
-                .padding(16.dp)
-                .weight(1f)
-        ) {
-            when (selectedMenuItem) {
-                "Personal Info" -> PersonalInfoTab()
-//                "Profile Picture" -> ProfilePictureTab()
-                "Body Measurements" -> BodyMeasurementsTab()
-                "Preferences" -> PreferencesTab()
             }
         }
     }
-}
-
-@SuppressLint("RememberReturnType")
-@Composable
-fun PersonalInfoTab() {
-    val context = LocalContext.current
-    val sharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
-    var name by remember { mutableStateOf(sharedPreferences.getString("name", "") ?: "") }
-    var gender by remember { mutableStateOf(sharedPreferences.getString("gender", "") ?: "") }
 
 
+    private fun signInWithGithub() {
+        val provider = OAuthProvider.newBuilder("github.com")
 
-        Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Color(0xFFADD8E6)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            // Display profile picture (placeholder image for now)
-            Image(
-                painter = painterResource(id = R.drawable.profile_user), // Use a drawable resource
-                contentDescription = "Profile Picture",
-                modifier = Modifier
-                    .height(100.dp)
-                    .width(100.dp)
-            )
+        // Optionally add additional scopes
+        provider.addCustomParameter("allow_signup", "false")
 
-            Spacer(modifier = Modifier.height(16.dp))
+        // Start the sign-in flow
+        auth.startActivityForSignInWithProvider(this, provider.build())
+            .addOnSuccessListener { authResult ->
+                val user = authResult.user
+                val email = user?.email ?: "No email"
+                val username = user?.displayName ?: email.substringBefore("@")
+                val githubId = user?.uid ?: "No ID"
 
-            // Display name information
-            Text(
-                text = "Name: $name",
-                style = MaterialTheme.typography.titleLarge,  // Updated text style ,
-                modifier = Modifier.fillMaxWidth(),
-                color = Color.Black
-            )
-
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-//
-            val gender = "Male"
-            Text(
-                text = "Gender: $gender",
-                style = MaterialTheme.typography.titleLarge,  // Use titleLarge instead of h6
-                modifier = Modifier.fillMaxWidth(),
-                color = Color.Black
-            )
-
-            Button(onClick = {
-                // Perform actions to navigate or edit profile info
-            }) {
-                Text(text = "Edit Profile Info")
-            }
-        }
-    }
-}
-
-@Composable
-fun BodyMeasurementsTab() {
-    val context = LocalContext.current
-    val sharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
-    var weight by remember { mutableStateOf(sharedPreferences.getString("weight", "") ?: "") }
-    var height by remember { mutableStateOf(sharedPreferences.getString("height", "") ?: "") }
-    var showMessage by remember { mutableStateOf(false) }
-
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Color(0xFFADD8E6)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            OutlinedTextField(
-                value = weight,
-                onValueChange = { weight = it },
-                label = { Text("Enter your weight (kg)") },
-                modifier = Modifier.fillMaxWidth(),
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            OutlinedTextField(
-                value = height,
-                onValueChange = { height = it },
-                label = { Text("Enter your height (cm)") },
-                modifier = Modifier.fillMaxWidth(),
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            Button(onClick = {
-                if (weight.isNotBlank() && height.isNotBlank()) {
-                    val editor = sharedPreferences.edit()
-                    editor.putString("weight", weight)
-                    editor.putString("height", height)
-                    editor.apply()
-                    showMessage = true
+                val intent = Intent(this, Landing::class.java).apply {
+                    putExtra("EMAIL", email)
+                    putExtra("USERNAME", username)
+                    putExtra("GITHUB_ID", githubId)
                 }
-            }) {
-                Text(text = "Save Measurements")
+                Toast.makeText(
+                    this,
+                    "Sign-in successful: ${authResult.user?.email}",
+                    Toast.LENGTH_SHORT
+                ).show()
+                startActivity(intent)
+                finish()
             }
-
-            if (showMessage) {
-                Text(
-                    text = "Measurements saved successfully!",
-                    color = Color.Green,
-                    modifier = Modifier.padding(top = 16.dp)
-                )
+            .addOnFailureListener { e ->
+                // Handle failure
+                Toast.makeText(this, "Sign-in failed: ${e.message}", Toast.LENGTH_SHORT).show()
             }
-        }
     }
-}
 
-@Composable
-fun PreferencesTab() {
-    // Placeholder for Preferences functionality
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = Color(0xFFADD8E6)
-    ) {
-        Column(
+
+    @Composable
+    fun GithubButton() {
+        Box(
             modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
+                .fillMaxSize() // Fill the entire screen
+                .padding(16.dp), // Optional padding around the box
+            contentAlignment = Alignment.Center // Center the content
         ) {
-            Text("Preferences Section")
-
-            // This is where you'd add preference settings like notifications, themes, etc.
-            Button(onClick = { /* Implement preferences logic */ }) {
-                Text(text = "Edit Preferences")
+            Button(onClick = { signInWithGithub() }) {
+                Text("Github Sign In")
             }
         }
     }
+
+    @Composable
+    fun GoogleSignInButton() {
+        val context = LocalContext.current
+        val coroutineScope = rememberCoroutineScope()
+
+        val onClick: () -> Unit = {
+
+            val credentialManager = CredentialManager.create(context)
+
+            val rawNonce = UUID.randomUUID().toString()
+            val bytes = rawNonce.toByteArray()
+            val md = MessageDigest.getInstance("SHA-256")
+            val digest = md.digest(bytes)
+            val hashedNonce = digest.fold("") { str, it -> str + "%02x".format(it) }
+
+            val googleIdOption: GetGoogleIdOption = GetGoogleIdOption.Builder()
+                .setFilterByAuthorizedAccounts(false)
+                .setServerClientId("44637818454-2fu80osaqikevbdfrl6s1cv06nfvo1ov.apps.googleusercontent.com")
+                .setNonce(hashedNonce)
+                .build()
+
+            val request: GetCredentialRequest = GetCredentialRequest.Builder()
+                .addCredentialOption(googleIdOption)
+
+                .build()
+            coroutineScope.launch {
+                try {
+
+                    val result = credentialManager.getCredential(
+                        request = request,
+                        context = context,
+                    )
+                    val credential = result.credential
+
+                    val googleIdTokenCredential = GoogleIdTokenCredential
+                        .createFrom(credential.data)
+                    val googleIdToken = googleIdTokenCredential.idToken
+
+
+
+                    val displayName = googleIdTokenCredential.displayName
+
+                    val intent = Intent(context, Landing::class.java).apply {
+                        putExtra("USERNAME", displayName)
+                    }
+                    Log.i("TAG", "User display name: $displayName")
+                    Log.i("TAG", googleIdToken)
+
+
+
+                    Toast.makeText(context, "Signed in", Toast.LENGTH_SHORT).show()
+
+
+                    context.startActivity(intent)
+                    finish()
+                } catch (e: GetCredentialException) {
+                    Toast.makeText(context, e.message, Toast.LENGTH_SHORT).show()
+                } catch (e: GoogleIdTokenParsingException) {
+                    Toast.makeText(context, e.message, Toast.LENGTH_SHORT).show()
+                }
+            }
+
+        }
+
+        Button(onClick = onClick) {
+            Text("Google")
+
+        }
+
+
+    }
+
+
 }
